fix Books n+1 queries w/@Formula annot on reservationCount, Deleted reservationIds list

+++ src/main/kotlin/ar/edu/unsam/phm/controller/BookController.kt
@@ -62,7 +62,7 @@ class BookController(
         @RequestParam userId: Long,
         @RequestParam pickUpDate: LocalDate,
         @RequestParam dropOffDate: LocalDate
-    ): Int =
+    ): Long =
         bookService.recalculateBibliokarmas(id, userId, pickUpDate, dropOffDate)
 
     @GetMapping("/book-genders")
