+++ src/main/kotlin/ar/edu/unsam/phm/services/ReservationService.kt
@@ -52,7 +52,7 @@ class ReservationService(
             throw BusinessException("Reserva no disponible en esa fecha")
         }
 
-        user.addBibliokarmas(book.calculateBibliokarmas(reservation.reservationDays(), user.bibliokarmas))
+        user.addBibliokarmas(book.calculateBibliokarmas(reservation.reservationDays(), user.bibliokarmas).toInt())
 
         userRepository.save(user)
         reservationRepository.save(reservation)
@@ -101,7 +101,7 @@ class ReservationService(
                 bibliokarmas = reservation.book.calculateBibliokarmas(
                     reservation.reservationDays(),
                     reservation.user.bibliokarmas
-                )
+                ).toInt()
             }
         }
     }
